Updated to 1.4.2 classes

+++ Source/Starbreeze/Private/SBZBlackMarketStoreItemButton.cpp
@@ -19,7 +19,7 @@ USBZInventoryBaseData* USBZBlackMarketStoreItemButton::GetItemData() const {
     return NULL;
 }
 
-void USBZBlackMarketStoreItemButton::GetItemCost(int32& Price, ESBZCurrencyCode& Currency) {
+void USBZBlackMarketStoreItemButton::GetItemCost(int64& Price, ESBZCurrencyCode& Currency) {
 }
